codegen: generate only requested class and remove unused config
Previous version was making lots of assumptions about relationships
between classes (statics, activation factories), and that was causing
problems when trying to generate some classes or interfaces. The
information about those relationships is stored in the CustomAttributes
table in the form of binary blobs, we should use that information to
generate the code without making any assumption.