codegen: add exclusive implemented interfaces when generating a class
Implemented interfaces may be exclusive to the runtime class, this means
that the interface is private, and should not be generated on its own.
This has forced us to change interface generation, because previously
all methods from exclusive interfaces were generated as static, and this
is not correct. Now we only do this for static and factory interfaces.
The process is manual, and there's probably some hint in the IDL to know
if a method should be called statically or not, but I could not find it.