store: move metadata file parser and context management to a new struct
Previously when generating a type, its context was stored in the
generator struct. But the generator was failing in some cases because
sometimes we need to get information about another type, and if that
type belongs to another metadata file, it would return a class not found
error when searching in the context. This commit fixes that by moving
the context management to a new metadata store abstraction, which holds
all contexts, and uses when looking for a type.