codegen: the delegate needs to know the IDD of the class it implements
The TypedEventHandler is a parametrized interface. But we need to know
the GUID of the specialized class (`TypedEventHandler<A>` does not have
the same GUID as `TypedEventHandler<B>`). And these GUIDs are not
included in the *.idl, so we cannot automatically generate them.
According to the docs "these interfaces are generated by the MIDL
compiler, which takes care to specialize each one, and it’s on these
specializations that it attaches the GUID representing the interface
identifier."