delegates: keep a goroutine alive during the lifecycle of delegates
There's an error in the Go runtime that causes a deadlock error whenever
we are waiting for a delegate callback to happen. This is because Go is
not able to detect that we are going to receive a syscall in the future.
See this for more information: https://github.com/golang/go/issues/55015