codegen/delegates: reuse callbacks to avoid reaching go limits
Only a limited number of callbacks may be created in a single Go
process, and any memory allocated for these callbacks is never released.
Between NewCallback and NewCallbackCDecl, at least 1024 callbacks can
always be created, but this may not be enough for certain applications.
This commit avoids that probly by changing delegates so that all of them
share the same four callbacks.

fixes #82