codegen: change static func names to avoid conflicts
This commit changes the naming strategy to include the class name as a
prefix of the static function names. Static functions were generated as
public functions using the method name found in the windows metadata.
Since all static functions of a namespace are added to the same Go
package, this could cause collisions when two classes had the same
static method (`FromBluetoothAddress` for example can be found in both
Bluetoothdevice and BluetoothLEDevice).

fixes #85