codegen: fix generic input parameters
Generic parameters are not yet supported so we use the unsafe.Pointer to
let the user pass any parameter it wants. We were not taking this into
account on function input parameters and we were wraping the
unsafe.Pointer value again in another unsafe.Pointer, which caused
errors for some users.

fixes #92